+++ src/api/hr.js
@@ -1,0 +1,6 @@
+import api from './axiosInstance';
+
+export const getHRData = async () => {
+  const response = await api.get('/hr');
+  return response.data;
+};