+++ src/api/inventory.js
@@ -1,0 +1,6 @@
+import axios from './axiosInstance';
+
+export const getInventoryData = async () => {
+  const response = await axios.get('/inventory');
+  return response.data;
+};