+++ src/api/legal.js
@@ -1,0 +1,6 @@
+import axios from './axiosInstance';
+
+export const getLegalData = async () => {
+  const response = await axios.get('/legal');
+  return response.data;
+};