+++ src/api/pkash.js
@@ -1,0 +1,6 @@
+import api from './axiosInstance';
+
+export const getPKashData = async () => {
+  const response = await api.get('/pkash');
+  return response.data;
+};