+++ src/api/purchases.js
@@ -1,0 +1,6 @@
+import api from './axiosInstance';
+
+export const getPurchasesData = async () => {
+  const response = await api.get('/purchases');
+  return response.data;
+};