+++ src/api/sales.js
@@ -1,0 +1,6 @@
+import axios from './axiosInstance';
+
+export const getSalesData = async () => {
+  const response = await axios.get('/sales');
+  return response.data;
+};